Update patch set 1
Patch Set 1:

Still kept the code handling the releases around in case we change our mind, or need to put out some other release later - so that this can get quickly picked up by a data sync.

Usually the way to test betas is to select "Beta" channel in the OTA updater, as the factory images would only be used for initial install.

I was actually looking at not showing beta builds when they're the same as stable which is when this clicked in - we don't need to show them at all, ever, really.

Patch-set: 1
Reviewer: Gerrit User 1000005 <1000005@35266860-6252-469c-a67d-185064479fc0>