Update patch set 2
Patch Set 2: Verified+1 Code-Review+2

as soon as there are more than one method of 'get', i will replace this get page with an index of options. but for now it is fine to get rid of the separate download.

Patch-set: 2
Reviewer: Gerrit User 1000005 <1000005@35266860-6252-469c-a67d-185064479fc0>
Label: Code-Review=+2
Label: Verified=+1