Location privacy: remove Fake Traveler
Fake Traveler is a *testing* tool and should not be recommended for privacy. It is easy to check this by installing it, setting it up, and testing if it works,
even if it does, get an app like WiFiAnalyzer, and it is clear that nearby Wi-Fi networks (and most likely Bluetooth beacons and cellular networks) are visible to apps, still. These can be easily used to determine the coordinates of the user.

Their repository also recommends turning off system location providers (in our case microG) as they may show the real location occasionally. Again, not a privacy feature.

Change-Id: Ib32da11b361194835d0865f4538727b7c4e3c6ef